PixivHelper.sanitize_filename(): correctly cutting filename on linux
Cutting filename to <= 249 bytes in UTF8 on Linux environment.
For fixing #525

For Linux, 255 bytes is the maximum for filename (not full path).
And almost all Linux environments use UTF-8 for path encoding.
And the path is used later with suffix ".pixiv" (consume 6 bytes).
So 249 bytes is maybe a minimum.
https://github.com/Nandaka/PixivUtil2/blob/dbb810716c4824724c03ce8e022a5b5d31f41800/PixivHelper.py#L590

Windows/Mac isn't affected.