Fix bug #1109 (#1110)
Added a workaround that performs a bit depth check for each image contained in the temporary ugoira processing folder. If some images have a lower bit depth number than others, these images are all encoded to have the same bit depth number corresponding to the highest among the images in the folder.