Fixes Byte Order Mark in Downloaded file list, again. (#1137)
codecs.open is stateless, so simple way doesn't work and we need to deal
with state, or we get a BOM on each newline. This is suboptimal and forces
more postprocessing of Downloaded file list for correct result.

All of this code could be deleted if irfan fixes it's unicode file bug.
Alas, bug persists in all text files irfan opens, including it's config.
So fix is unlikely. :^(

Note to self: test not only on one line, but on multiple lines.