Fix issue with fanbox coverImageUrl
It looks like the post list and posts themselves use a different key for the cover image (I am assuming this is new, due to the error only showing up now) so calling ParsePost on the list (which uses a "cover" object which either has a url key or nothing) will break as it is expecting simply a coverImageUrl key.  The change here will handle it differently depending on whether it finds the coverImageUrl key or not.