Update PixivModelFanbox.py
While gross, this change is likely necessary to prevent problems with an embedded post that has no cover image.  If someone can come up with a better check, please provide one.  If the cover URL behaves similarly to the post list, in the event that there is no cover image for a post, the "cover" object will be empty, so trying to retrieve the value for the "url" key will fail and crash PixivUtil2.  This change verifies that the "url" key exists before retrieving the value.