Fix issue with fanbox coverImageUrl (#1181) (#1185)
* Fix issue with fanbox coverImageUrl

It looks like the post list and posts themselves use a different key for the cover image (I am assuming this is new, due to the error only showing up now) so calling ParsePost on the list (which uses a "cover" object which either has a url key or nothing) will break as it is expecting simply a coverImageUrl key.  The change here will handle it differently depending on whether it finds the coverImageUrl key or not.

* Update content_provider.json

Apparently the embed json uses cover.url instead of coverImageUrl now, similarly to the post list.

* Update PixivModelFanbox.py

While gross, this change is likely necessary to prevent problems with an embedded post that has no cover image.  If someone can come up with a better check, please provide one.  If the cover URL behaves similarly to the post list, in the event that there is no cover image for a post, the "cover" object will be empty, so trying to retrieve the value for the "url" key will fail and crash PixivUtil2.  This change verifies that the "url" key exists before retrieving the value.

* Update fanbox_urlembed.json

Update test data with new cover url format

* Update PixivModelFanbox.py

Address situation where embedded post's cover object is None

* Update PixivModelFanbox.py

break instead of redundant assignment of root = None, since it's not longer necessary

* Further fixes for issue #1181

Address issue that arises when cover is a type other than "cover_image"